refactor(MarkRights): ; optimize logic
* avoid querying user groups for IP users
